[Silver III] Title: 주유소, Time: 340 ms, Memory: 35912 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -20,7 +20,7 @@ public class Main {
             cost[i] = Integer.parseInt(st.nextToken());
         }
 
-        int min = Integer.MAX_VALUE;
+        long min = Integer.MAX_VALUE;
         long result = 0;
         for(int i = 0; i < N-1; i++){
             min = Math.min(min, cost[i]);
